moved RandomNameProperty to _types
